Make sure wdb_cmd and tclcad_rt get correct data
Looks like tclcad_rt and wdb_cmd are the Tcl_CreateCommand cases where
we won't be passing in an MGED command structure as data.  Adding the
magic checks to be doubly sure we don't mess up and send the wrong
client data to the wrong functions through a chain of commands.