Update set, allow user data pointer in bu_structparse
Yeesh.  The combination of bu_structparse and shoehorning data through Tcl's
clientData mechanism is a real pain here.  Can't really use the same trick we
used with stdin_input - unlike that case, where we can free the containers just
prior to quit and and are supposed to use them until then, there doesn't seem
to be a logical place to hang the free operation here.

Instead, allow users to pass arbitrary data through a structparse entry with a
void data pointer - looks like we can thread the data needle that way, although
we had to update the initializers in our code to init the new parameter.